phase-1-agentic-navigation-planner/requirements.txt
# Phase 1: Agentic Navigation Planner
strands-agents
strands-tools
python-dotenv
aws-deepracer-control-v2
flask
